% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmzBoxes.R
\docType{data}
\name{AmzBoxes}
\alias{AmzBoxes}
\title{Length, width, and height measurements for 98 Amazon shipping boxes}
\format{
A data set of measurements for 98 Amazon shipping boxes. 
These data were downloaded from the  BoxDimensions website: (https://www.boxdimensions.com/).
The data set includes five variables:
\itemize{
   \item Amazon Box Size
   \item Length (inches) 
   \item Width (inches) 
   \item Height (inches)
   \item Volume (inches)
 }
}
\usage{
data(AmzBoxes)
}
\description{
Length, width, and height measurements for 98 Amazon shipping boxes
}
\examples{
data(AmzBoxes)

hist(AmzBoxes$`Length (inches)`,
     main = "Histogram of Box Lengths",
     xlab = "Length",
     col = "blue")


}
\keyword{datasets}
