\name{BadRBY}
\alias{BadRBY}
\title{Improper correlation matrix reported by Bentler and Yuan}
\description{Example improper R matrix reported by Bentler and Yuan (2011)}
\usage{ 
data(BadRBY)
}
\format{
{A 12 by 12 non-positive definite correlation matrix.}
}
\source{Bentler, P. M. & Yuan, K. H. (2011). Positive definiteness via off-diagonal scaling of a symmetric indefinite matrix. Psychometrika, 76(1), 119--123. 
}
\examples{
data(BadRBY)
}
\keyword{statistics}% at least one, from doc/KEYWORDS

