\name{BadRJN}
\alias{BadRJN}
\title{Improper R matrix reported by Joseph and Newman}
\description{Example NPD improper correlation matrix reported by Joseph and Newman}
\usage{ 
data(BadRJN)
}
\format{
{A 14 by 14 non-positive definite correlation matrix.}
}
\source{Joseph, D. L. & Newman, D. A. (2010). Emotional intelligence: an integrative meta-analysis and cascading model. Journal of Applied Psychology, 95(1), 54--78.}
\examples{
data(BadRJN)
}
\keyword{statistics}% at least one, from doc/KEYWORDS

