\name{BadRKtB}
\alias{BadRKtB}
\title{Improper R matrix reported by Knol and ten Berge}
\description{Example improper R matrix reported by Knol and ten Berge}
\usage{ 
data(BadRKtB)
}
\format{
{A 6 by 6 non-positive definite correlation matrix.}
}
\source{Knol, D. L. and Ten Berge, J. M. F.  (1989).  Least-squares approximation of an improper correlation matrix by a proper one.  
Psychometrika, 54(1), 53-61.  
}
\examples{
data(BadRKtB)
}
\keyword{statistics}% at least one, from doc/KEYWORDS

