\name{BadRLG}
\alias{BadRLG}
\title{Improper R matrix reported by Lurie and Goldberg}
\description{Example improper R matrix reported by Lurie and Goldberg}
\usage{ 
data(BadRLG)
}
\format{
{A 3 by 3 non-positive definite correlation matrix.}
}
\source{Lurie, P. M. & Goldberg, M. S. (1998). An approximate method for sampling correlated random variables from partially-specified distributions. Management Science, 44(2), 203--218. 
}
\examples{
data(BadRLG)
}
\keyword{statistics}% at least one, from doc/KEYWORDS

