% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fungible-package.R
\name{BadRRM}
\alias{BadRRM}
\title{Improper R matrix reported by Rousseeuw and Molenberghs}
\format{A 3 by 3 non-positive definite correlation matrix.}
\source{
Rousseeuw, P. J. & Molenberghs, G. (1993). Transformation of non
positive semidefinite correlation matrices. Communications in
Statistics--Theory and Methods, 22(4), 965--984.
}
\description{
Example improper R matrix reported by Rousseeuw and Molenberghs
}
\examples{

data(BadRRM)

}
\keyword{statistics}
