% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Box20.R
\docType{data}
\name{Box20}
\alias{Box20}
\title{Length, width, and height measurements for Thurstone's 20 boxes}
\format{A data set of measurements for Thurstone's 20 hypothetical boxes. 
The data set includes three variables:
\itemize{
   \item \strong{x} Box length
   \item \strong{y} Box width
   \item \strong{z} Box height
 }}
\usage{
data(Box20)
}
\description{
Length, width, and height measurements for Thurstone's 20 hypothetical boxes
}
\examples{
data(Box20)

hist(Box20$x,
     main = "Histogram of Box Lengths",
     xlab = "Length",
     col = "blue")


}
\keyword{datasets}
