% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb.R
\name{cb}
\alias{cb}
\title{Cudeck & Browne (1992) model error method}
\usage{
cb(mod, target_rmsea)
}
\arguments{
\item{mod}{A `fungible::simFA()` model object.}

\item{target_rmsea}{(scalar) Target RMSEA value.}
}
\description{
Generate a population correlation matrix using the model described in Cudeck
and Browne (1992). This function uses the implementation of the Cudeck and
Browne method from Ken Kelley's MBESS package.
}
\references{
Cudeck, R., & Browne, M. W. (1992). Constructing a covariance
  matrix that yields a specified minimizer and a specified minimum
  discrepancy function value. *Psychometrika*, *57*(3), 357–369.
  <https://doi.org/10/cq6ckd>

  Kelley, K. (2017). MBESS (Version 4.0.0 and higher)
  [computer software and manual]. Accessible from http://cran.r-project.org.
}
