% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2r.r
\name{d2r}
\alias{d2r}
\title{Convert Degrees to Radians}
\usage{
d2r(deg)
}
\arguments{
\item{deg}{Angle in degrees.}
}
\value{
Angle in radians.
}
\description{
A simple function to convert degrees to radians
}
\examples{

d2r(90)

}
\keyword{Statistics}
