% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fals.R
\name{fals}
\alias{fals}
\title{Unweighted least squares factor analysis}
\usage{
fals(R, nfactors, TreatHeywood = TRUE)
}
\arguments{
\item{R}{Input correlation matrix.}

\item{nfactors}{Number of factors to extract.}

\item{TreatHeywood}{If TreatHeywood = TRUE then a penalized least squares
function is used to bound the commonality estimates below 1.0.
Default(TreatHeywood = TRUE).}
}
\value{
\item{loadings}{Unrotated factor loadings. If a Heywood case is
present in the initial solution then the model is re-estimated via
non-iterated principal axes with max(rij^2) as fixed communaility (h2)
estimates.} \item{h2}{Vector of final commonality estimates.}
\item{uniqueness}{Vector of factor uniquenesses, i.e. (1 - h2).}
\item{Heywood}{(logical) TRUE if a Heywood case was produced in the LS
solution.} \item{TreatHeywood}{(logical) Value of the TreatHeywood
argument.} \item{converged}{(logical) TRUE if all values of the gradient are
sufficiently close to zero.} \item{MaxAbsGrad}{The maximum absolute value of
the gradient at the solution.}  
\item{f.value}{The discrepancy value associated with the final solution.}
}
\description{
Unweighted least squares factor analysis
}
\examples{


Rbig <- fungible::rcor(120)                   
out1 <- fals(R = Rbig, 
             nfactors = 2,
             TreatHeywood = TRUE)

}
\seealso{
Other Factor Analysis Routines:
\code{\link[=BiFAD]{BiFAD()}},
\code{\link{Box26}},
\code{\link[=GenerateBoxData]{GenerateBoxData()}},
\code{\link[=Ledermann]{Ledermann()}},
\code{\link[=SLi]{SLi()}},
\code{\link[=SchmidLeiman]{SchmidLeiman()}},
\code{\link[=faAlign]{faAlign()}},
\code{\link[=faEKC]{faEKC()}},
\code{\link[=faIB]{faIB()}},
\code{\link[=faLocalMin]{faLocalMin()}},
\code{\link[=faMB]{faMB()}},
\code{\link[=faMain]{faMain()}},
\code{\link[=faScores]{faScores()}},
\code{\link[=faSort]{faSort()}},
\code{\link[=faStandardize]{faStandardize()}},
\code{\link[=faX]{faX()}},
\code{\link[=fapa]{fapa()}},
\code{\link[=fareg]{fareg()}},
\code{\link[=fsIndeterminacy]{fsIndeterminacy()}},
\code{\link[=orderFactors]{orderFactors()}},
\code{\link[=print.faMB]{print.faMB()}},
\code{\link[=print.faMain]{print.faMain()}},
\code{\link[=promaxQ]{promaxQ()}},
\code{\link[=summary.faMB]{summary.faMB()}},
\code{\link[=summary.faMain]{summary.faMain()}}
}
\author{
Niels Waller
}
\concept{Factor Analysis Routines}
\keyword{Statistics}
