% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW.R
\docType{data}
\name{HW}
\alias{HW}
\title{Six data sets that  yield a Heywood case}
\format{Each data set is a  matrix with 150 rows and 12 variables:
\describe{ Each data set (HW1, HW2, ... HW6) represents a hypothetical sample
of 150 subjects from a population 3-factor model.  
The population factor loadings are given in \code{HW$popLoadings}.
}}
\usage{
data(HW)
}
\description{
Six data sets that  yield a Heywood case in a 3-factor model.
}
\examples{
data(HW)

# Compute a principal axis factor analysis 
# on the first data set  
RHW <- cor(HW$HW1)  
fapaOut <- faMain(R = RHW, 
                 numFactors = 3, 
                 facMethod = "fapa", 
                 rotate = "oblimin",
                 faControl = list(treatHeywood = FALSE))


fapaOut$faFit$Heywood
round(fapaOut$h2, 2)


}
\keyword{datasets}
