\name{plot.monte}
\alias{plot.monte}
\title{Plot Method for Class Monte}
\description{
plot method for class "monte"
}
\usage{
\method{plot}{monte}(x, \dots)
}
\arguments{
\item{x}{ An object of class 'monte', usually, a result of a call to \code{monte}.}
\item{\dots}{Optional arguments passed to plotting function.}
}
\value{
 The function \code{plot.monte} creates a scatter plot of matrices plot (a splom plot).  Cluster membership is denoted by different colors in the plot. 
}
\examples{
#plot(monte.object)
}


