% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.faMain.R
\name{print.faMain}
\alias{print.faMain}
\title{Print  Method for an Object of Class faMain}
\usage{
\method{print}{faMain}(x, ..., digits = 2, Set = 1, itemSort = FALSE)
}
\arguments{
\item{x}{(Object of class \strong{faMain}) The returned object 
from a call to \strong{faMain}.}

\item{\dots}{Additional arguments affecting the summary produced.}

\item{digits}{(Integer) Print output with user-specified number of significant digits. Default \code{digits = 2}.}

\item{Set}{\itemize{
 \item{integer} (Integer) Summarize the solution from the specified solution set. 
 \item{'UnSpun'} (Character) Summarize the solution from the rotated output that was 
  produced by rotating from the unrotated (i.e., unspun) factor orientation.
}}

\item{itemSort}{(Logical) If TRUE, sort the order of the observed variables to produce
a "staircase"-like pattern. In bifactor models (i.e., bifactorT and bifactorQ) item 
sorting is determined by  the magnitudes of the group factor loadings.
Defaults to \code{itemSort} = FALSE.}
}
\description{
Print  Method for an Object of Class faMain
}
\seealso{
Other Factor Analysis Routines:
\code{\link[=BiFAD]{BiFAD()}},
\code{\link{Box26}},
\code{\link[=GenerateBoxData]{GenerateBoxData()}},
\code{\link[=Ledermann]{Ledermann()}},
\code{\link[=SLi]{SLi()}},
\code{\link[=SchmidLeiman]{SchmidLeiman()}},
\code{\link[=faAlign]{faAlign()}},
\code{\link[=faEKC]{faEKC()}},
\code{\link[=faIB]{faIB()}},
\code{\link[=faLocalMin]{faLocalMin()}},
\code{\link[=faMB]{faMB()}},
\code{\link[=faMain]{faMain()}},
\code{\link[=faScores]{faScores()}},
\code{\link[=faSort]{faSort()}},
\code{\link[=faStandardize]{faStandardize()}},
\code{\link[=faX]{faX()}},
\code{\link[=fals]{fals()}},
\code{\link[=fapa]{fapa()}},
\code{\link[=fareg]{fareg()}},
\code{\link[=fsIndeterminacy]{fsIndeterminacy()}},
\code{\link[=orderFactors]{orderFactors()}},
\code{\link[=print.faMB]{print.faMB()}},
\code{\link[=promaxQ]{promaxQ()}},
\code{\link[=summary.faMB]{summary.faMB()}},
\code{\link[=summary.faMain]{summary.faMain()}}
}
\concept{Factor Analysis Routines}
