\name{r2d}
\alias{r2d}
\title{Convert Radians to Degrees}
\description{
Convert radian measure to degrees. 
}
\usage{
r2d(radian)
}
\arguments{
 \item{radian}{Radian measure of an angle}
}
\value{
 \item{}{Degree measure of an angle} 
}
\examples{
 r2d(.5*pi)
}
\keyword{Statistics}% at least one, from doc/KEYWORDS
