% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmad.R
\name{Rmad}
\alias{Rmad}
\title{Generate a random R matrix with given MAD from a user-defined, population R}
\usage{
Rmad(Rpop, MAD = 0.03, tries = 1e+05, SEED = NULL)
}
\arguments{
\item{Rpop}{(matrix) A model-implied, population correlation matrix.}

\item{MAD}{(scalar) The desired mean absolute deviation value (i.e., 
\eqn{mad(Rpop - R) = MAD} where \eqn{mad} computes the mean absolute deviation 
for the nonredundant, non-diagonal values of its argument).}

\item{tries}{(scalar) The maximum number of tries for the optimization algorithm. 
Default(\eqn{tries = 1E5}).}

\item{SEED}{(numeric)  The starting seed for the random number generator. 
If SEED is not supplied then the program will generate (and return) a randomly
generated seed.}
}
\value{
\itemize{
 \item \strong{R} A random \strong{R} matrix with a known MAD from \strong{Rpop}.
 \item \strong{Rpop} The model-implied, population correlation matrix.
 \item \strong{Delta} The vector of model-errors.
 \item \strong{H}  A hollow matrix used to create a fungible \strong{R} matrix.
 \item \strong{iter} Iteration number at convergence (or max iteration number).
 \item \strong{MAD} The target mean absolute deviation.
 \item \strong{converged}. A Boolean that describes the convergence status of 
 the optimization algorithm.
 \item  \strong{SEED} The initial value for the random number generator.
}
}
\description{
Rmad(Rpop, MAD = .10, tries = 1E5, SEED = NULL)
}
\examples{
 # Example 1
  Rpop <- matrix(.35, 6, 6)
  diag(Rpop) <- 1
  
  out <- Rmad(Rpop, MAD = .038, SEED = 123)
               
  out$R |> round(3)            
 
# E = mattix of model error values
  E = Rpop - out$R 
  mean(abs(E[lower.tri(E, diag = FALSE)]))
   
}
\author{
Niels Waller
}
