% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rplus.R
\name{Rplus}
\alias{Rplus}
\title{Generate a random R matrix with all rij > 0}
\usage{
Rplus(Seed, NVar = NULL, rdist = NULL, param1 = NULL, param2 = NULL, s = NULL)
}
\arguments{
\item{Seed}{(scalar) A seed for the random number generator.  If 
\code{Seed = NULL} then the program will generate a random seed.}

\item{NVar}{(integer)  The order of the generated correlation (\strong{R}) matrix.}

\item{rdist}{(character) A one or two letter character string that determines
the distribution for the elements of the Cholesky factor of the
randomly generated covariance matrix. Possible values include
\itemize{
 \item \strong{B} Beta(param1, param2)
 \item \strong{X} Chisq(param1
 \item \strong{E} Exponential(param1)
 \item \strong{F} F(param1, param2)
 \item \strong{G} Gamma(param1, param2)
 \item \strong{IB} Inverse Beta(param1, param2)
 \item \strong{IX} Inverse Chisq(param1)
 \item \strong{LN} Log Normal(param1, param2)
 \item \strong{U} Uniform(param1, param2)
 \item  \strong{W} Weibull(param1, param2)
}}

\item{param1}{(scalar). The first parameter of \strong{rdist}.}

\item{param2}{(scalar). The second parameter of \strong{rdist}}

\item{s}{(scalar). An optional scalar \eqn{s \in [0,1]} that
scales the generated correlations by a factor of \strong{s}. 
Default (\code{s = NULL}). If \code{s} is supplied then no 
correlations in \strong{R} will be larger than \code{s}.}
}
\value{
\itemize{
 \item \strong{R}  The generated \strong{R} matrix.  
 \item \strong{NVar} The order of \strong{R}.
 \item \strong{rdist} The user-supplied character strong for \strong{rdist},
 \item \strong{param1} The first parameter of \strong{rdist}. 
 \item \strong{param2} The possible second parameter of \strong{rdist}.
 \item \strong{s} The user-supplied scaling factor.
 \item \strong{Seed} the starting value for the random number generator.
}
}
\description{
Rplus(Seed, NVar = NULL,  rdist =  NULL,  param1 = NULL,  
      param2 = NULL, s = NULL)
}
\examples{
 # --- Example 1: Generate two random R with positive rij ----
 R = Rplus(Seed = 123,  
           NVar = 6,  
           rdist =  "F",  
           param1 = 2.5,  
           param2 = 25,
           s = .8)$R

 R |> round(2)

R = Rplus(Seed = 456,  
          NVar = 6,  
          rdist =  "F",  
          param1 = 2.5,  
          param2 = 25,
          s = .8)$R
       
 R |> round(2)   

}
\author{
Niels G. Waller
}
