% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semify.R
\name{semify}
\alias{semify}
\title{Generate an sem model from a simFA model object}
\usage{
semify(mod)
}
\arguments{
\item{mod}{A `fungible::simFA()` model object.}
}
\description{
Generate an sem model from a simFA model object
}
\examples{
ex_mod <- fungible::simFA(Seed = 42)
semify(mod = ex_mod)
}
