% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fungible-package.R
\name{BadRBY}
\alias{BadRBY}
\title{Improper correlation matrix reported by Bentler and Yuan}
\format{
A 12 by 12 non-positive definite correlation matrix.
}
\source{
Bentler, P. M. & Yuan, K. H. (2011). Positive definiteness via
off-diagonal scaling of a symmetric indefinite matrix. Psychometrika, 76(1),
119--123.
}
\description{
Example improper R matrix reported by Bentler and Yuan (2011)
}
\examples{

data(BadRBY)

}
\keyword{statistics}
