% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fungible-package.R
\name{BadRKtB}
\alias{BadRKtB}
\title{Improper R matrix reported by Knol and ten Berge}
\format{A 6 by 6 non-positive definite correlation matrix.}
\source{
Knol, D. L. and Ten Berge, J. M. F.  (1989).  Least-squares
approximation of an improper correlation matrix by a proper one.
Psychometrika, 54(1), 53-61.
}
\description{
Example improper R matrix reported by Knol and ten Berge
}
\examples{

data(BadRKtB)

}
\keyword{statistics}
