% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fungible-package.R
\name{BadRLG}
\alias{BadRLG}
\title{Improper R matrix reported by Lurie and Goldberg}
\format{
A 3 by 3 non-positive definite correlation matrix.
}
\source{
Lurie, P. M. & Goldberg, M. S. (1998). An approximate method for
sampling correlated random variables from partially-specified distributions.
Management Science, 44(2), 203--218.
}
\description{
Example improper R matrix reported by Lurie and Goldberg
}
\examples{

data(BadRLG)

}
\keyword{statistics}
