\name{BadRRM}
\alias{BadRRM}
\title{Improper R matrix reported by Rousseeuw and Molenberghs}
\description{Example improper R matrix reported by Rousseeuw and Molenberghs}
\usage{ 
data(BadRRM)
}
\format{
{A 3 by 3 non-positive definite correlation matrix.}
}
\source{Rousseeuw, P. J. & Molenberghs, G. (1993). Transformation of non positive semidefinite correlation matrices. Communications in Statistics--Theory and Methods, 22(4), 965--984.  
}
\examples{
data(BadRRM)
}
\keyword{statistics}% at least one, from doc/KEYWORDS

