% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.monte.R
\name{plot.monte}
\alias{plot.monte}
\title{Plot Method for Class Monte}
\usage{
\method{plot}{monte}(x, ...)
}
\arguments{
\item{x}{An object of class 'monte', usually, a result of a call to
\code{monte}.}

\item{\dots}{Optional arguments passed to plotting function.}
}
\value{
The function \code{plot.monte} creates a scatter plot of matrices
plot (a splom plot).  Cluster membership is denoted by different colors in
the plot.
}
\description{
plot method for class "monte"
}
\examples{

#plot(monte.object)

}
