% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.faMain.R
\name{print.faMain}
\alias{print.faMain}
\title{Print  Method for an Object of Class faMain}
\usage{
\method{print}{faMain}(x, ..., digits = 2, Set = 1, itemSort = FALSE)
}
\arguments{
\item{x}{(Object of class \strong{faMain}) The returned object 
from a call to \strong{faMain}.}

\item{\dots}{Additional arguments affecting the summary produced.}

\item{digits}{(Integer) Print output with user-specified number of significant digits. Default \code{digits = 2}.}

\item{Set}{\itemize{
 \item{integer} (Integer) Summarize the solution from the specified solution set. 
 \item{'UnSpun'} (Character) Summarize the solution from the rotated output that was 
  produced by rotating from the unrotated (i.e., unspun) factor orientation.
}}

\item{itemSort}{(Logical) If TRUE, sort the order of the observed variables to produce
a "staircase"-like pattern. In bifactor models (i.e., bifactorT and bifactorQ) item 
sorting is determined by  the magnitudes of the group factor loadings.
Defaults to \code{itemSort} = FALSE.}
}
\description{
Print  Method for an Object of Class faMain
}
\seealso{
Other Factor Analysis Routines: 
\code{\link{BiFAD}()},
\code{\link{Box26}},
\code{\link{GenerateBoxData}()},
\code{\link{Ledermann}()},
\code{\link{SLi}()},
\code{\link{SchmidLeiman}()},
\code{\link{faAlign}()},
\code{\link{faEKC}()},
\code{\link{faIB}()},
\code{\link{faLocalMin}()},
\code{\link{faMB}()},
\code{\link{faMain}()},
\code{\link{faScores}()},
\code{\link{faSort}()},
\code{\link{faStandardize}()},
\code{\link{faX}()},
\code{\link{fals}()},
\code{\link{fapa}()},
\code{\link{fareg}()},
\code{\link{orderFactors}()},
\code{\link{print.faMB}()},
\code{\link{promaxQ}()},
\code{\link{summary.faMB}()},
\code{\link{summary.faMain}()}
}
\concept{Factor Analysis Routines}
