% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2d.r
\name{r2d}
\alias{r2d}
\title{Convert Radians to Degrees}
\usage{
r2d(radian)
}
\arguments{
\item{radian}{Radian measure of an angle.}
}
\value{
Degree measure of an angle.
}
\description{
Convert radian measure to degrees.
}
\examples{

 r2d(.5*pi)

}
\keyword{Statistics}
