% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcos.r
\name{vcos}
\alias{vcos}
\title{Compute the Cosine Between Two Vectors}
\usage{
vcos(x, y)
}
\arguments{
\item{x}{A p x 1 vector.}

\item{y}{A p x 1 vector.}
}
\value{
\item{Cosine between x and y}{}
}
\description{
Compute the cosine between two vectors.
}
\examples{

x <- rnorm(5)
y <- rnorm(5)
vcos(x, y)

}
\keyword{Statistics}
