% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnorm.r
\name{vnorm}
\alias{vnorm}
\title{Norm a Vector to Unit Length}
\usage{
vnorm(x)
}
\arguments{
\item{x}{An n by 1 vector.}
}
\value{
the scaled (i.e., unit length) input vector
}
\description{
Norm a vector to unit length.
}
\examples{

 x <- rnorm(5)
 v <- vnorm(x)
 print(v)

}
\author{
Niels Waller
}
\keyword{statistics}
